Revert "Add a changePassword test"
This reverts commit 7e519e11b7b8376d92fdec2492732e75abcda350.

This commit added a test that upgraded from srp to bcrypt via
changePassword for a user with only an email address, in addition to the
test that existed for a user with a username. Having two such tests is
now silly because Meteor.changePassword no longer has different code
paths for email/username users.